organized api endpoints into subfolders
